Modal code re-organization
