+++ GravatarApp/CommonViews/MainMenu/MainMenuOptions.swift
@@ -44,53 +44,7 @@ struct MainMenuOptions: View {
     }
 }
 
-struct AboutView: View {
-    @EnvironmentObject var modalManager: ModalPresentationManager
 
-    var body: some View {
-        VStack(alignment: .leading) {
-            title("About Gravatar")
-            Text("v\(getAppVersion())")
-                .foregroundStyle(.secondary)
-            title("Get help")
-            link("support.gravatar.com", url: "https://support.gravatar.com")
-            link("support@gravatar.com", url: "mailto:support@gravatar.com")
-            title("Legal")
-            link("Terms of Service", url: "https://gravatar.com")
-            link("Privacy Policy", url: "https://gravatar.com")
-
-            Button {
-                modalManager.dismiss()
-            } label: { Text("Done").frame(maxWidth: .infinity) }
-                .buttonStyle(.actionButton(style: .secondary))
-                .padding(.top, 16)
-        }
-        .padding()
-    }
-
-    private func title(_ text: String) -> some View {
-        Text(text)
-            .font(.title3)
-            .fontWeight(.semibold)
-            .padding(.top, 16)
-    }
-
-    private func link(_ text: String, url: String) -> some View {
-        Link(text, destination: URL(string: url)!)
-            .foregroundStyle(.secondary)
-    }
-
-    func getAppVersion() -> String {
-        if let appVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String {
-            return appVersion
-        }
-        return "?"
-    }
-}
-
-#Preview {
-    AboutView()
-}
 
 private enum Localized {
     static let visitProfileTitle: String = NSLocalizedString(
